aura_aegis_sim/requirements.txt
streamlit
pandas
plotly
numpy
scikit-learn
torch --index-url https://download.pytorch.org/whl/cpu
joblib
cryptography
pyserial
